The entire chain of connection from StarterClass to WS and response back. I still need to test for multiple WS points.

+++ Try50/src/aos/AddInterface.java
@@ -2,5 +2,5 @@ package aos;
 
 public interface AddInterface {
 	public int add();
-	public double myload();
+	public int myload();
 }
